Move tests inline
